Restructured dnsendpoint controller
